Fixed mapping and print value from controller

+++ gui/MainWindow.py
@@ -38,7 +38,7 @@ class MainWindow(QMainWindow):
         self.ui.graphWidget.ci.layout.setSpacing(20)
 
         width = QApplication.primaryScreen().size().width()
-        self.ui.splitterHorizontal.setSizes([width/8, width*5/8, width*2/8])
+        self.ui.splitterHorizontal.setSizes([int(width/8), int(width*5/8), int(width*2/8)])
 
     def on_buttonClearLog_clicked(self):
         self.ui.teLog.clear()
